requirements.txt
python-telegram-bot==20.8
aiohttp
aiosqlite
nest_asyncio
python-dotenv